scripts/gcp/worker/requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-cloud-bigquery==3.34.0
